refactor(engine): move classes to streamprocessor package
